Skip test

+++ korlibs-io/test/korlibs/io/vfs/LocalVfsTest.kt
@@ -86,6 +86,8 @@ class LocalVfsTest {
 
     @Test
     fun testUnixPermissions() = suspendTest({ (Platform.isJvm && Platform.isUnix) || Platform.isMac || Platform.isLinux }) {
+        if (Platform.isJsDenoJs) return@suspendTest
+
         val chmod = "0713".toInt(8)
         val file = tempVfs["korio-temp123.bin"]
         file.delete()
